fix some code in the reduction functions

+++ cpu_funcs.h
@@ -8,16 +8,16 @@
 
 void initiate_program(int pid, int num_processes);
 
-void fill_hash(double* weights, int pid);
+void fill_hash(float* weights, int pid);
 void print_hash();
 
-double find_best_mutant_cpu(int pid, ProgramData* data, Mutant* return_mutant, int first_offset, int last_offset);
+float find_best_mutant_cpu(int pid, ProgramData* data, Mutant* return_mutant, int first_offset, int last_offset);
 
 ProgramData* read_seq_and_weights_from_file(FILE* file, ProgramData* data);
-int write_results_to_file(FILE* file, char* mutant, int offset, double score);
+int write_results_to_file(FILE* file, char* mutant, int offset, float score);
 
 void pretty_print(ProgramData* data, char* mut, int offset, int char_offset);
-double get_score_and_signs(char* seq1, char* seq2, double* weights, int offset, char* signs);
+float get_score_and_signs(char* seq1, char* seq2, float* weights, int offset, char* signs);
 void print_with_offset(char* chrs, int offset, int char_offset);
 
 #endif //__CPU_FUNCS_H__
